commit by chh

+++ order-api/src/main/java/com/honghao/cloud/orderapi/domain/entity/WaybillBcList.java
@@ -61,7 +61,7 @@ public class WaybillBcList {
 
     private BigDecimal receiveLatitude;
 
-    private Byte isTimely;
+    private Integer isTimely;
 
     private Date etStartTime;
 
@@ -71,11 +71,11 @@ public class WaybillBcList {
 
     private String cargoType;
 
-    private Byte orderStatus;
+    private Integer orderStatus;
 
-    private Byte orderPayment;
+    private Integer orderPayment;
 
-    private Byte orderPayStatus;
+    private Integer orderPayStatus;
 
     private BigDecimal deliveryFee;
 
@@ -83,13 +83,13 @@ public class WaybillBcList {
 
     private String orderRemark;
 
-    private Byte orderSource;
+    private Integer orderSource;
 
-    private Byte orderType;
+    private Integer orderType;
 
-    private Byte singleWay;
+    private Integer singleWay;
 
-    private Byte deliveryMode;
+    private Integer deliveryMode;
 
     private Date orderTime;
 
@@ -101,5 +101,5 @@ public class WaybillBcList {
 
     private Date updateTime;
 
-    private Byte deleteFlag;
+    private Integer deleteFlag;
 }
